// packages/integration-oura/src/oura-api.d.ts
export interface components {
    schemas: {
// ...
        DailyActivityModel: {
            /** Id */
            id: string;
            /**
             * Class 5 Min
             * @description 5-minute activity classification for the activity period:
             *     * ```0```	non wear
             *     * ```1``` rest
             *     * ```2``` inactive
             *     * ```3``` low activity
             *     * ```4``` medium activity
             *     * ```5``` high activity
             */
            class_5_min: string | null;
            /**
             * Score
             * @description Activity score in range ```[1, 100]```
             */
            score: number | null;
            /**
             * Active Calories
             * @description Active calories expended (in kilocalories)
             */
            active_calories: number;
            /**
             * Average Met Minutes
             * @description Average metabolic equivalent (MET) in minutes
             */
            average_met_minutes: number;
            contributors: components["schemas"]["ActivityContributors"];
            /**
             * Equivalent Walking Distance
             * @description Equivalent walking distance (in meters) of energy expenditure
             */
            equivalent_walking_distance: number;
            /**
             * High Activity Met Minutes
             * @description High activity metabolic equivalent (MET) in minutes
             */
            high_activity_met_minutes: number;
            /**
             * High Activity Time
             * @description High activity metabolic equivalent (MET) in seconds
             */
            high_activity_time: number;
            /**
             * Inactivity Alerts
             * @description Number of inactivity alerts received
             */
            inactivity_alerts: number;
            /**
             * Low Activity Met Minutes
             * @description Low activity metabolic equivalent (MET) in minutes
             */
            low_activity_met_minutes: number;
            /**
             * Low Activity Time
             * @description Low activity metabolic equivalent (MET) in seconds
             */
            low_activity_time: number;
            /**
             * Medium Activity Met Minutes
             * @description Medium activity metabolic equivalent (MET) in minutes
             */
            medium_activity_met_minutes: number;
            /**
             * Medium Activity Time
             * @description Medium activity metabolic equivalent (MET) in seconds
             */
            medium_activity_time: number;
            met: components["schemas"]["SampleModel"];
            /**
             * Meters To Target
             * @description Remaining meters to target (from ```target_meters```
             */
            meters_to_target: number;
            /**
             * Non Wear Time
             * @description The time (in seconds) in which the ring was not worn
             */
            non_wear_time: number;
            /**
             * Resting Time
             * @description Resting time (in seconds)
             */
            resting_time: number;
            /**
             * Sedentary Met Minutes
             * @description Sedentary metabolic equivalent (MET) in minutes
             */
            sedentary_met_minutes: number;
            /**
             * Sedentary Time
             * @description Sedentary metabolic equivalent (MET) in seconds
             */
            sedentary_time: number;
            /**
             * Steps
             * @description Total number of steps taken
             */
            steps: number;
            /**
             * Target Calories
             * @description Daily activity target (in kilocalories)
             */
            target_calories: number;
            /**
             * Target Meters
             * @description Daily activity target (in meters)
             */
            target_meters: number;
            /**
             * Total Calories
             * @description Total calories expended (in kilocalories)
             */
            total_calories: number;
            /**
             * Day
             * Format: date
             * @description The ```YYYY-MM-DD``` formatted local date indicating when the daily activity occurred
             */
            day: string;
            /** @description ISO 8601 formatted local timestamp indicating the start datetime of when the daily activity occurred */
            timestamp: components["schemas"]["LocalDateTime"];
        };
// ...
        DailyReadinessModel: {
            /** Id */
            id: string;
            /** @description Contributors of the daily readiness score. */
            contributors: components["schemas"]["ReadinessContributors"];
            /**
             * Day
             * Format: date
             * @description Day that the daily readiness belongs to.
             */
            day: string;
            /**
             * Score
             * @description Daily readiness score.
             */
            score: number | null;
            /**
             * Temperature Deviation
             * @description Temperature deviation in degrees Celsius.
             */
            temperature_deviation: number | null;
            /**
             * Temperature Trend Deviation
             * @description Temperature trend deviation in degrees Celsius.
             */
            temperature_trend_deviation: number | null;
            /** @description Timestamp of the daily readiness. */
            timestamp: components["schemas"]["LocalDateTime"];
        };
// ...
        DailySleepModel: {
            /** Id */
            id: string;
            /** @description Contributors for the daily sleep score. */
            contributors: components["schemas"]["SleepContributors"];
            /**
             * Day
             * Format: date
             * @description Day that the daily sleep belongs to.
             */
            day: string;
            /**
             * Score
             * @description Daily sleep score.
             */
            score: number | null;
            /** @description Timestamp of the daily sleep. */
            timestamp: components["schemas"]["LocalDateTime"];
        };
        /** DailySpO2AggregatedValuesModel */
        DailySpO2AggregatedValuesModel: {
            /**
             * Average
             * @description Average oxygen saturation (SpO2) throughout the night.
             */
            average: number;
        };
        /** DailySpO2Model */
        DailySpO2Model: {
            /** Id */
            id: string;
            /**
             * Day
             * Format: date
             */
            day: string;
            /** @description The SpO2 percentage value aggregated over a single day. */
            spo2_percentage: components["schemas"]["DailySpO2AggregatedValuesModel"] | null;
            /**
             * Breathing Disturbance Index
             * @description Breathing Disturbance Index (BDI) calculated using detected SpO2 drops from timeseries. Values should be in range [0, 100]
             */
            breathing_disturbance_index: number | null;
        };
// ...
        HeartRateModel: {
            /** Bpm */
            bpm: number;
            source: components["schemas"]["HeartRateSource"];
            timestamp: components["schemas"]["LocalDateTime"];
        };
// ...
        PublicWorkout: {
            /** @description Unique identifier of the object. */
            id: string;
            /** @description Type of the workout activity. */
            activity: string;
            /** @description Energy burned in kilocalories during the workout. */
            calories?: number | null;
            /** @description Day when the workout occurred. */
            day: components["schemas"]["ISODate"];
            /** @description Distance traveled in meters during the workout. */
            distance?: number | null;
            /** @description Timestamp indicating when the workout ended. */
            end_datetime: components["schemas"]["LocalizedDateTime"];
            /** @description Intensity of the workout. */
            intensity: components["schemas"]["PublicWorkoutIntensity"];
            /** @description User-defined label for the workout. */
            label?: string | null;
            /** @description Possible workout sources. */
            source: components["schemas"]["PublicWorkoutSource"];
            /** @description Timestamp indicating when the workout started. */
            start_datetime: components["schemas"]["LocalizedDateTime"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}